git ignore

+++ src/main/java/com/example/demo/dtos/RequestOrderCardDTO.java
@@ -16,6 +16,6 @@ public class RequestOrderCardDTO {
     List<CardOrderDTO> cardInfo;
     String paymentMethodCode;
     UserEntity userInfo;
-    Long totalPrice;
+    float totalPrice;
     String ip_address;
 }
