+++ src/main/java/com/example/demo/repositories/tables/entities/PaymentEntity.java
@@ -34,7 +34,7 @@ public class PaymentEntity implements Serializable {
     String createUser;
 
     @Column(name = "PRICE")
-    Long price;
+    float price;
 
     @Column(name = "ORDER_ID")
     String orderId;
